% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richcast-package.R
\docType{package}
\name{richcast-package}
\alias{richcast}
\alias{richcast-package}
\title{richcast: Hindcast Species Distributions and Assemblage Richness Through Time}
\description{
Fits a species distribution model for every species near a region, projects it onto palaeoclimate reconstructions across a series of time slices, and stacks the results into assemblage richness surfaces. Ranges come from IUCN Red List polygons supplied by the user; each model is an equally weighted ensemble of a random forest and MaxEnt, trained on pseudo-presences sampled from the range polygon and evaluated by AUC and the continuous Boyce index. Climate data are retrieved via 'pastclim'. A trait table for rodents from Ecke et al. (2022) \doi{10.1038/s41467-022-35273-7} is bundled as a worked example.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nxmarom/richcast}
  \item Report bugs at \url{https://github.com/nxmarom/richcast/issues}
}

}
\author{
\strong{Maintainer}: Nimrod Marom \email{nxmarom@gmail.com}

Authors:
\itemize{
  \item Nimrod Marom \email{nxmarom@gmail.com}
}

}
\keyword{internal}
