% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{richness_surface}
\alias{richness_surface}
\title{Extract a richness surface from a series}
\usage{
richness_surface(series, time)
}
\arguments{
\item{series}{A \code{richcast_series} from \code{\link[=run_hindcast_series]{run_hindcast_series()}}.}

\item{time}{Year CE, or \code{"present"}.}
}
\value{
A \code{SpatRaster} of species counts.
}
\description{
Extract a richness surface from a series
}
\seealso{
\code{\link[=richness_grid]{richness_grid()}}, \code{\link[=run_hindcast_series]{run_hindcast_series()}}
}
