% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.R
\name{pastclim_climate}
\alias{pastclim_climate}
\title{Climate source: live pastclim queries}
\usage{
pastclim_climate(
  dataset_present = "WorldClim_2.1_5m",
  dataset_past = "CHELSA_trace21k_1.0_0.5m_vsi",
  present_time = 1985,
  times = NULL
)
}
\arguments{
\item{dataset_present, dataset_past}{pastclim dataset names.}

\item{present_time}{Year CE treated as "present" when fitting.}

\item{times}{Numeric vector of years CE this source can serve.}
}
\value{
A \code{richcast_climate}.
}
\description{
Retrieves slices through \pkg{pastclim} on demand. Convenient for
exploration; for a full run over many species and time slices, prepare the
slices once with \code{\link[=prepare_climate]{prepare_climate()}} and use \code{\link[=climate_dir]{climate_dir()}} instead, since
every species would otherwise re-stream the same rasters.
}
\examples{
clim <- pastclim_climate(times = seq(850, 1850, by = 100))
clim
}
\references{
Leonardi, M., Hallet, E. Y., Beyer, R., Krapp, M., & Manica,
A. (2023). pastclim 1.2: an R package to easily access and use
paleoclimatic reconstructions. \emph{Ecography}, 2023, e06481.
\doi{10.1111/ecog.06481}

Cite the reconstruction you use as well, as \pkg{pastclim} asks:

Beyer, R. M., Krapp, M., & Manica, A. (2020). High-resolution terrestrial
climate, bioclimate and vegetation for the last 120,000 years.
\emph{Scientific Data}, 7, 236. \doi{10.1038/s41597-020-0552-1}

Karger, D. N., Nobis, M. P., Normand, S., Graham, C. H., & Zimmermann,
N. E. (2023). CHELSA-TraCE21k -- high-resolution (1 km) downscaled
transient temperature and precipitation data since the Last Glacial
Maximum. \emph{Climate of the Past}, 19, 439-456.
\doi{10.5194/cp-19-439-2023}

Fick, S. E., & Hijmans, R. J. (2017). WorldClim 2: new 1-km spatial
resolution climate surfaces for global land areas. \emph{International
Journal of Climatology}, 37(12), 4302-4315. \doi{10.1002/joc.5086}
}
