% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richcast-package.R
\docType{package}
\name{richcast-package}
\alias{richcast}
\alias{richcast-package}
\title{richcast: Hindcast Species Distributions and Assemblage Richness Through Time}
\description{
A taxon-agnostic pipeline for projecting species distribution models onto palaeoclimate reconstructions and summarising the resulting assemblage richness over a series of time slices. Range data are supplied by the user from any of several backends (IUCN Red List shapefiles, GBIF occurrences, or arbitrary 'sf' polygons); climate data are retrieved via 'pastclim'. A trait table for rodents from Ecke et al. (2022) \doi{10.1038/s41467-022-35273-7} is bundled as a worked example.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nxmarom/richcast}
  \item Report bugs at \url{https://github.com/nxmarom/richcast/issues}
}

}
\author{
\strong{Maintainer}: Nimrod Marom \email{nxmarom@gmail.com}

Authors:
\itemize{
  \item Nimrod Marom \email{nxmarom@gmail.com}
}

}
\keyword{internal}
